requirements.txt
nonebot2>=2.2.0
nonebot-adapter-onebot>=2.0.0
httpx
fastapi
pydantic
fuzzyfinder
requests
jieba
jinja2
nonebot-plugin-htmlrender
pyppeteer
tencentcloud-sdk-python>=3.0.580
nonebot-plugin-apscheduler
openai
nonebot-plugin-tortoise-orm
tortoise-orm
